Users and Edit task and create

+++ src/main/java/schedule/pro/application/Entity/Dto/CreateInboxMessageDto.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class CreateInboxMessageDto {
     public String message;
-    public long userId;
+    public Integer userId;
 }
